+++ CHANGELOG.md
@@ -20,6 +20,8 @@ the following form: YYYY.0M.0D.
 
 - For Part 2 I made a source java file where I showed how NailAppointment component would look like and is very simple
 - Designed kernel and enhanced interfaces for NailAppointment component
+- Designed abstract class for NailAppointment component
+- Implemented toString, equals, and hashCode for NailAppointment
 
 
 ## [2025.09.19]